### 2020-09-08 #2650 @whedon
[PRE REVIEW]: pd-parser: A tool for Matching Photodiode Deflection Events to Time-Stamped Events
**Submitting author:** @alexrockhill (<a href="http://orcid.org/0000-0003-3868-7453">Alexander Rockhill</a>)
**Repository:** <a href="https://github.com/alexrockhill/pd-parser" target ="_blank">https://github.com/alexrockhill/pd-parser</a>
**Version:** 0.1
