### 2020-09-08 #2650 @danielskatz
👋 @alexrockhill - please add something to the README that explains how this is research software.  What kinds of researchers need this software? (this is in the paper)

Also, note that some images in the README seem broken
