+++ README.rst
@@ -3,19 +3,19 @@
 |icon| pd-parser
 ================
 
-.. |icon| image:: ./docsrc/favicon.png
-  :width: 50
+.. |icon| image:: ./favicon.png
+  :width: 25
 
 A tool to parse photodiode events from a possibly corrupted channel, compatible with BIDS formatting.
 
 Corrupted data may look like so:
 
-.. image:: ./docsrc/figs/excluded_events.png
+.. image:: ./figs/excluded_events.png
    :width: 800
 
 Which may lead to some excluded events and differences between the events and the expected timing based on the behavior that look like so:
 
-.. image:: ./docsrc/figs/event_diffs.png
+.. image:: ./figs/event_diffs.png
    :width: 400
 
 ``pd-parser`` matches deflection events on a potentially corrupted photodiode channel with time-stamped events generated by the computer that changes the luminance sensed by the photodiode. This tool is for research that needs high-performance timing synchronization, especially in clinical settings where having a channel record a photodiode is the best or only method of task-integration available and where the photodiode is especially likely to be corrupted by unavoidable environmental noise. First, photodiode time series data is extracted from an electrophysiology data file. The photodiode data can be on a single channel or two channels that are then bipolar re-referenced. Candidate photodiode events from this photodiode time series are identified based on matching a square-wave template. Time-stamped events from the computer triggering changes to the luminance of the photodiode are read from a tab-separated value (tsv) file, and the best alignment of these events relative to photodiode deflection events is then found while accounting for any drift between computer clocks of separate recording devices. Events are excluded where when the difference between the photodiode event and the time-stamped event are greater then a specified threshold. This discrepancy in timing can occur when a monitor doesn't update the display for one or more frames most often because the computer paused execution of the program controlling luminance to do background tasks. Additional events can then be added using the time of each event relative to the corresponding photodiode-synchronized event. Adding relative events may be needed for any number of reasons. For instance, events may occur too rapidly in series relative to the monitor refresh rate to have a photodiode deflection for each event, or the photodiode deflection may effect neighboring channels on the amplifier with its large deflection and so be placed synchronous with an event, such the onset of fixation, where it will not effect other, more important events. Finally, the raw data and events data can be saved in brain imaging data structure (BIDS) format, which allows the behavioral events to be stored in a standardized format without modifying the underlying raw electrophysiology file.
