### 2020-09-11 8dab222
added tests for CLI and fixed bugs with CLI, added array option for seconds on in simulation
