+++ pd_parser/commands/run.py
@@ -9,6 +9,7 @@ import pd_parser
 
 def find_pd_params():
     """Plot the photodiode channel to find parameters for the parser."""
+    import matplotlib.pyplot as plt
     parser = argparse.ArgumentParser()
     parser.add_argument('fname', type=str,
                         help='The electrophysiology filepath')
@@ -24,6 +25,7 @@ def find_pd_params():
     args = parser.parse_args()
     pd_parser.find_pd_params(args.fname, pd_ch_names=args.pd_ch_names,
                              verbose=args.verbose)
+    plt.show()
 
 
 def parse_pd():
@@ -49,6 +51,14 @@ def parse_pd():
     parser.add_argument('--exclude_shift', type=float, required=False,
                         default=0.05, help='How many seconds off to exclude a '
                         'photodiode-behavioral event difference')
+    parser.add_argument('--resync', type=float, required=False,
+                        default=0.075, help='How large of a difference '
+                        'to use to resynchronize events. This is for when '
+                        'events are off but not by much and so they should '
+                        'be excluded but are still needed to fit an alignment.'
+                        'Increase if the alignment is failing because too '
+                        'many events are being excluded, decrease to speed up '
+                        'execution.')
     parser.add_argument('--chunk', type=float, required=False,
                         default=2, help='How large to window the '
                         'photodiode events, should >> 2 * longest event. '
@@ -79,14 +89,11 @@ def parse_pd():
                         'photodiode event. Probably don\'t change but '
                         'increasing will reduce false-positives and '
                         'decreasing will reduce false-negatives.')
-    parser.add_argument('--resync', type=float, required=False,
-                        default=0.075, help='How large of a difference '
-                        'to use to resynchronize events. This is for when '
-                        'events are off but not by much and so they should '
-                        'be excluded but are still needed to fit an alignment.'
-                        'Increase if the alignment is failing because too '
-                        'many events are being excluded, decrease to speed up '
-                        'execution.')
+    parser.add_argument('--add_events', type=bool, required=False,
+                        default=False, help='Whether to run the parser '
+                        'a second time to add more events from '
+                        'deflections corresponding to multiple events '
+                        'on the same channel')
     parser.add_argument('--verbose', default=True, type=bool,
                         required=False,
                         help='Set verbose output to True or False.')
@@ -96,8 +103,42 @@ def parse_pd():
     pd_parser.parse_pd(
         args.fname, pd_event_name=args.pd_event_name, behf=args.behf,
         beh_col=args.beh_col, pd_ch_names=args.pd_ch_names, chunk=args.chunk,
-        baseline=args.baseline, overlap=args.overlap,
-        exclude_shift=args.exclude_shift, zscore=args.zscore, min_i=args.min_i,
+        exclude_shift=args.exclude_shift, resync=args.resync,
+        zscore=args.zscore, min_i=args.min_i, baseline=args.baseline,
+        add_events=args.add_events, verbose=args.verbose,
+        overwrite=args.overwrite)
+
+
+def add_pd_off_event():
+    """Run add_pd_off command."""
+    parser = argparse.ArgumentParser()
+    parser.add_argument('fname', type=str,
+                        help='The electrophysiology filepath')
+    parser.add_argument('--off_event_name', type=str, required=False,
+                        default='StimOff',
+                        help='The name of the photodiode event')
+    parser.add_argument('--chunk', type=float, required=False,
+                        default=2, help='The same chunk as used for '
+                        '`parse_pd`.')
+    parser.add_argument('--zscore', type=float, required=False,
[…]
