+++ pd_parser/parse_pd.py
@@ -160,11 +160,12 @@ def _find_pd_candidates(pd, chunk_i, baseline_i, zscore, min_i, verbose=True):
                                     range(-min_i, min_i + 1) for this_pd_c in
                                     pd_candidates.values()]):
                             pd_candidates[name].add(i + e)
-    pd_candidates = (pd_candidates['down'] if
-                     len(pd_candidates['down']) > len(pd_candidates['up']) else
-                     pd_candidates['up'])
+    pd_direction = 'down' if \
+        len(pd_candidates['down']) > len(pd_candidates['up']) else 'up'
+    pd_candidates = pd_candidates[pd_direction]
     if verbose:
-        print(f'{len(pd_candidates)} photodiode candidate events found')
+        print(f'{len(pd_candidates)} {pd_direction}-deflection photodiode '
+              'candidate events found')
     sorted_pds = np.array(sorted(pd_candidates))
     return pd_candidates, sorted_pds
 
@@ -527,8 +528,9 @@ def find_pd_params(fname, pd_ch_names=None, verbose=True):
 
 
 def parse_pd(fname, pd_event_name='Fixation', behf=None,
-             beh_col='fix_onset_time', pd_ch_names=None, exclude_shift=0.03,
-             chunk=2, zscore=20, min_i=10, baseline=0.25, resync=0.075,
+             beh_col='fix_onset_time', pd_ch_names=None,
+             exclude_shift=0.03, resync=0.075,
+             chunk=2, zscore=20, min_i=10, baseline=0.25,
              add_events=False, overwrite=False, verbose=True):
     """Parse photodiode events.
 
@@ -556,6 +558,18 @@ def parse_pd(fname, pd_event_name='Fixation', behf=None,
     exclude_shift: float
         How many seconds different than expected from the behavior events
         to exclude that event. Use `find_pd_params` to determine if unsure.
+    resync : float
+        The number of seconds to difference allowed to still use a photodiode
+        event to resynchronize with time-stamped events. Events with
+        differences between `resync` and `exclude_shift` will still be
+        used for alignment but will be excluded from the events. When
+        `exclude_shift` is smaller than `resync`, this parameter allows
+        event differences less than `exclude_shift` to be removed without
+        losing an alignment which depends on resynchronizing to these events
+        between `exclude_shift` and `resync`. This is most likely to happen
+        when the drift between behavior events and the photodiode is large,
+        so many events are to be excluded for being off by a small amount
+        but still correctly correspond to a behavior event.
     chunk: float
         The size of the window to chunk the photodiode events by
         should be larger than 2x the longest photodiode event.
@@ -570,18 +584,6 @@ def parse_pd(fname, pd_event_name='Fixation', behf=None,
         How much relative to the chunk to use to idenify the time before
         the photodiode event. This should not be changed most likely
         unless there is a specific reason/issue.
-    resync : float
-        The number of seconds to difference allowed to still use a photodiode
-        event to resynchronize with time-stamped events. Events with
-        differences between `resync` and `exclude_shift` will still be
-        used for alignment but will be excluded from the events. When
-        `exclude_shift` is smaller than `resync`, this parameter allows
-        event differences less than `exclude_shift` to be removed without
-        losing an alignment which depends on resynchronizing to these events
-        between `exclude_shift` and `resync`. This is most likely to happen
[…]

[…]
