### 2020-09-11 #2650 @sappelhoff
According to the [JOSS conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) I'd have such a conflict for reviewing this piece, because I am working with Alex on [mne-bids](https://github.com/mne-tools/mne-bids/), which we published a paper on in 2019.

I think I could still make a relatively impartial judgment (this is up to @oliviaguest to decide) ... but if we can find another reviewer that'd be good as well.
