### 2020-09-11 #2650 @danielskatz
My preference would be to avoid the appearance of conflicts unless we don't have a good alternative, which it sounds like we do, in this case.
