### 2020-09-11 #2650 @libertyh
Happy to do it, but I've never reviewed for JOSS before, so let me know if there are any general guidelines for how to do this and I'm happy to.
