### 2020-09-11 #2650 @alexrockhill
The documentation for reviewing is here https://joss.readthedocs.io/en/latest/reviewer_guidelines.html :)
