### 2020-09-11 #2650 @libertyh
One very minor potential conflict of interest is that Nicole Swann and I were on this paper together (https://www.nature.com/articles/s41597-019-0105-7), but this is a large list of authors (N=39) with a large number of iEEG contributors, and I have otherwise not worked with these authors directly. But just for transparency's sake I thought I'd mention, though it does rule out many iEEG/ECoG people if used in this way.
