### 2020-09-11 #2650 @danielskatz
From the point of view of JOSS, this is not a conflict that should cause a problem
