### 2020-09-11 #2650 @oliviaguest
OK, cool! @libertyh do you want to review this given the above (no conflict from JOSS' perspective)?
