### 2020-09-11 #2650 @sneakers-the-rat
re: COI, Alex and I are both at University of Oregon (though I believe in different departments), I've done some light troubleshooting of a server with him but haven't collaborated on any major work. I think I can be impartial but also know that doesn't quite work for COI.

let me know if that is ok for JOSS
