### 2020-09-11 #2650 @danielskatz
Well
> In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution.

So this is a conflict for that reason.  Again, if we can find a reviewer without conflicts, that would be better.  If not, I think this is ok.
