### 2020-09-11 a70783a
added find off events, fixed some bugs
