+++ pd_parser/tests/test_parse_pd.py
@@ -20,8 +20,9 @@ import mne
 from mne.utils import _TempDir, run_subprocess
 
 import pd_parser
-from pd_parser.parse_pd import (_read_tsv, _to_tsv,
-                                _read_raw, _load_beh_df, _get_pd_data,
+from pd_parser.parse_pd import (_read_tsv, _to_tsv, _read_raw,
+                                _load_beh_df, _get_pd_channel_data,
+                                _get_pd_data, _check_if_pd_event,
                                 _find_pd_candidates, _pd_event_dist,
                                 _check_alignment, _find_best_alignment,
                                 _exclude_ambiguous_events,
@@ -47,14 +48,16 @@ raw_tmp.info['dig'] = None
 raw_tmp.info['line_freq'] = 60
 
 pd_event_name = 'Fixation'
+off_event_name = 'Stim Off'
 beh_col = 'fix_onset_time'
 pd_ch_names = ['pd']
 exclude_shift = 0.03
-chunk = 2
-zscore = 20
+max_len = 1
+zscore = 100
 min_i = 10
 baseline = 0.25
 resync = 0.075
+verbose = True
 
 
 # from mne_bids.tests.test_write._bids_validate
@@ -78,6 +81,7 @@ def _bids_validate():
     return _validate
 
 
+@pytest.mark.filterwarnings('ignore::RuntimeWarning')
 def test_core():
     out_dir = _TempDir()
     # test tsv
@@ -103,28 +107,36 @@ def test_core():
     with pytest.raises(ValueError, match='Unable to write'):
         _to_tsv('foo.bar', dict(test=1))
     # test read
-    raw, events = pd_parser.simulate_pd_data(n_secs_on=1.)
+    raw, beh_df, events, corrupted_indices = pd_parser.simulate_pd_data()
+    pd = raw._data[0]
     raw.save(op.join(out_dir, 'test-raw.fif'), overwrite=True)
     with pytest.raises(ValueError, match='not recognized'):
         _read_raw('foo.bar')
     raw2 = _read_raw(op.join(out_dir, 'test-raw.fif'))
-    np.testing.assert_array_almost_equal(raw._data, raw2._data, decimal=5)
+    np.testing.assert_array_almost_equal(raw._data, raw2._data, decimal=4)
     # test load beh
     with pytest.raises(ValueError, match='not in the columns'):
         _load_beh_df(op.join(basepath, 'pd_events.tsv'), 'foo')
     # test get pd data
-    raw, events = pd_parser.simulate_pd_data()
-    with pytest.raises(ValueError, match='not in raw channel names'):
+    with pytest.raises(ValueError, match='in raw channel names'):
         _get_pd_data(raw, ['foo'])
+    with pytest.raises(ValueError, match='in raw channel names'):
+        _get_pd_channel_data(raw, ['foo'])
     # test find pd candidates
     exclude_shift_i = np.round(raw.info['sfreq'] * exclude_shift).astype(int)
-    chunk_i = np.round(raw.info['sfreq'] * chunk).astype(int)
-    baseline_i = np.round(chunk_i * baseline / 2).astype(int)
+    max_len_i = np.round(raw.info['sfreq'] * max_len).astype(int)
+    baseline_i = np.round(max_len_i * baseline / 2).astype(int)
     resync_i = np.round(raw.info['sfreq'] * resync).astype(int)
+    assert _check_if_pd_event(pd, events[0, 0] - baseline_i,
+                              max_len_i, zscore, min_i, baseline_i) == \
+        ('up', events[0, 0], events[0, 0] + raw.info['sfreq'])  # one sec event
+    assert _check_if_pd_event(pd, baseline_i,
+                              max_len_i, zscore, min_i, baseline_i) == \
+        (None, None, None)
[…]
