+++ pd_parser/parse_pd.py
@@ -95,14 +95,27 @@ def _load_beh_df(behf, beh_col):
     return df[beh_col], df
 
 
+def _get_pd_channel_data(raw, pd_ch_names):
+    """Get the time-series data from the channel names."""
+    if any([ch not in raw.ch_names for ch in pd_ch_names]):
+        raise ValueError(f'Not all pd_ch_names, {pd_ch_names}, '
+                         'in raw channel names')
+    if len(pd_ch_names) == 2:
+        pd = raw._data[raw.ch_names.index(pd_ch_names[0])]
+        pd -= raw._data[raw.ch_names.index(pd_ch_names[1])]
+    else:
+        pd = raw._data[raw.ch_names.index(pd_ch_names[0])]
+    pd -= np.median(pd)
+    return pd
+
+
 def _get_pd_data(raw, pd_ch_names):
     """Get the names of the photodiode channels from the user."""
     # if pd_ch_names provided
     if pd_ch_names is not None:
-        for ch in pd_ch_names:
-            if ch not in raw.ch_names:
-                raise ValueError(f'{ch} in pd_ch_names {pd_ch_names} not '
-                                 'in raw channel names')
+        if any([ch not in raw.ch_names for ch in pd_ch_names]):
+            raise ValueError(f'Not all pd_ch_names, {pd_ch_names}, '
+                             'in raw channel names')
     else:  # if no pd_ch_names provided
         pd_ch_names = input('Enter photodiode channel names separated by a '
                             'comma or type "plot" to plot the data first:\t')
@@ -123,43 +136,52 @@ def _get_pd_data(raw, pd_ch_names):
                       '2 names for bipolar reference')
         pd_ch_names = [ch.strip() for ch in pd_ch_names.split(',')]
     # get pd data using channel names
-    if len(pd_ch_names) == 2:
-        pd = raw._data[raw.ch_names.index(pd_ch_names[0])]
-        pd -= raw._data[raw.ch_names.index(pd_ch_names[1])]
-    else:
-        pd = raw._data[raw.ch_names.index(pd_ch_names[0])]
-    pd -= np.median(pd)
+    pd = _get_pd_channel_data(raw, pd_ch_names)
     return pd, pd_ch_names
 
 
-def _find_pd_candidates(pd, chunk_i, baseline_i, zscore, min_i, verbose=True):
+def _check_if_pd_event(pd, i, max_len_i, zscore, min_i, baseline_i):
+    """Take one stretch of data and determine if there is an event there.
+
+    Move in chunks twice as long as your longest photodiode signal
+    with the first 0.25 as baseline to test whether the signal goes
+    above/below baseline and back below/above.
+    The event onset must have a minimum length `min_i`.
+    """
+    b = pd[i - baseline_i:i]
+    # use twice the length so event can be centered
+    s = (pd[i:i + 2 * max_len_i] - np.median(b)) / np.std(b)
+    binned_s = np.digitize(s, [-np.inf, -zscore, zscore, np.inf]) - 2
+    for direction, binary_s in {'up': binned_s, 'down': s < -binned_s}.items():
+        onset = np.where(binary_s[:max_len_i] == 1)[0]
+        if onset.size > min_i:  # must be long enough
+            e = onset[0]
+            if all(binary_s[:e] == 0):  # must start off:
+                # must have an offset and no more events
+                offset = np.where(binary_s[e:] < 1)[0]
+                if offset.size > 0:
+                    oe = offset[0]
+                    if oe <= max_len_i and \
+                            all(binary_s[e + oe:e + max_len_i] == 0):
+                        return direction, i + e, i + e + oe
+    return None, None, None
+
+
+def _find_pd_candidates(pd, max_len_i, baseline_i, zscore, min_i,
[…]
