+++ pd_parser/__init__.py
@@ -4,6 +4,6 @@ __version__ = '0.2.dev0'
 
 
 from pd_parser.parse_pd import (find_pd_params, parse_pd,  # noqa
-                                add_pd_relative_events,  # noqa
+                                add_pd_off_event, add_pd_relative_events,  # noqa
                                 add_pd_events_to_raw, pd_parser_save_to_bids,
                                 simulate_pd_data)  # noqa
