+++ docs/_sources/whats_new.rst.txt
@@ -22,6 +22,7 @@ Current
 
 Changelog
 ~~~~~~~~~
+- Fixed core parsing algorithm to be more strict about an uncorrupted baseline; previously the baseline could have noise > zscore in the opposite direction of the deflection.
 - Added vertical lines to plots to show the difference between expected and actual events.
 
 
@@ -32,7 +33,7 @@ Bug
 
 API
 ~~~
-- Added ``off_event`` as a parameter for :func:`pd_parser.parse_pd` to make an event when the photodiode turns off.
+- Added :func:`pd_parser.add_pd_off_event` to make an event when the photodiode turns off.
 - Added ``n_secs_on`` as an array in :func:`pd_parser.simulate_pd_data`.
 - Added ``resync`` parameter to :func:`pd_parser.parse_pd` so that events could be used to sychronize but then excluded.
 

[…]
