+++ docs/generated/cli.rst
@@ -1,209 +1,0 @@
-:orphan:
-
-.. _python_cli:
-
-======================================
-pd_parser Command Line Interface (CLI)
-======================================
-
-Here we list the pd_parser tools that you can use from the command line.
-
-.. contents:: Contents
-   :local:
-   :depth: 1
-
-
-
-.. _gen_add_pd_events_to_raw:
-
-====================
-
-.. rst-class:: callout
-
---------------------
-add_pd_events_to_raw
---------------------
-                            [--overwrite OVERWRITE]
-usage: add_pd_events_to_raw [-h] [--out_fname OUT_FNAME] [--verbose VERBOSE]
-
-positional arguments:
-  fname                 The electrophysiology filepath
-
-optional arguments:
-  -h, --help            show this help message and exit
-  --out_fname OUT_FNAME
-                        The name to save out the new raw file out to
-  --verbose VERBOSE     Set verbose output to True or False.
-  --overwrite OVERWRITE
-                        Whether to overwrite
-
-
-
-.. _gen_add_pd_relative_events:
-
-======================
-
-.. rst-class:: callout
-
-----------------------
-add_pd_relative_events
-----------------------
-                              [--relative_event_cols [RELATIVE_EVENT_COLS [RELATIVE_EVENT_COLS ...]]]
-usage: add_pd_relative_events [-h] [--behf BEHF]
-                              [--verbose VERBOSE] [--overwrite OVERWRITE]
-                              fname
-
-positional arguments:
-  fname                 The electrophysiology filepath
-
-optional arguments:
-  -h, --help            show this help message and exit
-  --behf BEHF           The behavioral tsv filepath
-  --relative_event_cols [RELATIVE_EVENT_COLS [RELATIVE_EVENT_COLS ...]]
-                        A behavioral column in the tsv file that has the time
-                        relative to the photodiode events on the same trial as
-                        in the `--beh_col event.
-  --relative_event_names [RELATIVE_EVENT_NAMES [RELATIVE_EVENT_NAMES ...]]
-                        The name of the corresponding `--relative_event_cols`
-                        events
-  --verbose VERBOSE     Set verbose output to True or False.
-  --overwrite OVERWRITE
-                        Whether to overwrite
-
-
-
-.. _gen_find_pd_params:
-
-==============
-
-.. rst-class:: callout
-
---------------
-find_pd_params
---------------
-                      [--verbose VERBOSE]
-usage: find_pd_params [-h] [--pd_ch_names [PD_CH_NAMES [PD_CH_NAMES ...]]]
-
-positional arguments:
-  fname                 The electrophysiology filepath
-
-optional arguments:
-  -h, --help            show this help message and exit
-  --pd_ch_names [PD_CH_NAMES [PD_CH_NAMES ...]]
-                        The name(s) of the channels with the photodiode data.
-                        Can be one channel for common referenced recording or
-                        two for a bipolar recording. If not provided, the data
-                        will be plotted for the user to pick
-  --verbose VERBOSE     Set verbose output to True or False.
-
-
-
-.. _gen_parse_pd:
-
-========
-
-.. rst-class:: callout
-
---------
-parse_pd
---------
-                [--beh_col BEH_COL]
-usage: parse_pd [-h] [--pd_event_name PD_EVENT_NAME] [--behf BEHF]
-                [--exclude_shift EXCLUDE_SHIFT] [--chunk CHUNK]
-                [--zscore ZSCORE] [--min_i MIN_I]
-                [--alignment_prop ALIGNMENT_PROP] [--baseline BASELINE]
-                [--overlap OVERLAP] [--verbose VERBOSE]
-                [--overwrite OVERWRITE]
-                fname
-
-positional arguments:
[…]
