+++ docs/_sources/pd_parser.parse_pd.rst.txt
@@ -1,6 +1,0 @@
-﻿pd\_parser.parse\_pd
-====================
-
-.. currentmodule:: pd_parser
-
-.. autofunction:: parse_pd
