+++ docs/_sources/pd_parser.add_pd_events_to_raw.rst.txt
@@ -1,6 +1,0 @@
-﻿pd\_parser.add\_pd\_events\_to\_raw
-===================================
-
-.. currentmodule:: pd_parser
-
-.. autofunction:: add_pd_events_to_raw
