+++ docs/_sources/pd_parser.add_pd_relative_events.rst.txt
@@ -1,6 +1,0 @@
-﻿pd\_parser.add\_pd\_relative\_events
-====================================
-
-.. currentmodule:: pd_parser
-
-.. autofunction:: add_pd_relative_events
