+++ docs/_sources/generated/pd_parser.find_pd_params.rst.txt
@@ -1,6 +1,0 @@
-﻿pd\_parser.find\_pd\_params
-===========================
-
-.. currentmodule:: pd_parser
-
-.. autofunction:: find_pd_params
