+++ docsrc/generated/cli.rst
@@ -50,3 +50,244 @@ optional arguments
 
 
 
+
+
+.. _gen_add_pd_off_events:
+
+add_pd_off_events
+=================
+
+.. rst-class:: callout
+
+usage: ``add_pd_off_events fname [-h] [--off_event_name OFF_EVENT_NAME] [--max_len MAX_LEN] [--zscore ZSCORE] [--min_i MIN_I] [--baseline BASELINE] [--verbose VERBOSE] [--overwrite OVERWRITE]``
+
+positional arguments
+--------------------
+fname
+	The electrophysiology filepath
+
+
+
+optional arguments
+------------------
+-h, --help		show this help message and exit
+
+
+--off_event_name OFF_EVENT_NAME		The name of the photodiode event
+
+
+--max_len MAX_LEN		The length of the longest photodiode event
+
+
+--zscore ZSCORE		The same zscore as used for `parse_pd`.
+
+
+--min_i MIN_I		The same min_i as used for `parse_pd`.
+
+
+--baseline BASELINE		The same baseline as used for `parse_pd`.
+
+
+--verbose VERBOSE		Set verbose output to True or False.
+
+
+--overwrite OVERWRITE		Whether to overwrite
+
+
+
+
+
+
+
+.. _gen_add_pd_relative_events:
+
+add_pd_relative_events
+======================
+
+.. rst-class:: callout
+
+usage: ``add_pd_relative_events fname [-h] [--behf BEHF] [--relative_event_cols LIST_OF_RELATIVE_EVENT_COLS] [--relative_event_names LIST_OF_RELATIVE_EVENT_NAMES] [--verbose VERBOSE] [--overwrite OVERWRITE]``
+
+positional arguments
+--------------------
+fname
+	The electrophysiology filepath
+
+
+
+optional arguments
+------------------
+-h, --help		show this help message and exit
+
+
+--behf BEHF		The behavioral tsv filepath
+
+
+--relative_event_cols LIST_OF_RELATIVE_EVENT_COLS		A behavioral column in the tsv file that has the time relative to the photodiode events on the same trial as in the `beh_col` event.
+
+
+--relative_event_names LIST_OF_RELATIVE_EVENT_NAMES		The name of the corresponding `relative_event_cols` events
+
+
+--verbose VERBOSE		Set verbose output to True or False.
+
+
+--overwrite OVERWRITE		Whether to overwrite
+
+
+
+
+
+
+
+.. _gen_find_pd_params:
+
+find_pd_params
+==============
+
+.. rst-class:: callout
+
+usage: ``find_pd_params fname [-h] [--pd_ch_names LIST_OF_PD_CH_NAMES] [--verbose VERBOSE]``
+
+positional arguments
+--------------------
+fname
+	The electrophysiology filepath
+
+
+
+optional arguments
+------------------
+-h, --help		show this help message and exit
+
+
+--pd_ch_names LIST_OF_PD_CH_NAMES		The name(s) of the channels with the photodiode data. Can be one channel for common referenced recording or two for a bipolar recording. If not provided, the data will be plotted for the user to pick
+
+
+--verbose VERBOSE		Set verbose output to True or False.
+
+
+
+
+
+
+
+.. _gen_parse_pd:
+
+parse_pd
+========
+
+.. rst-class:: callout
+
[…]
