### 2020-09-11 e2d7717
format bullet docstring change
