### 2020-09-15 #2650 @libertyh
I can still review if you need me! Thanks!

> On Sep 11, 2020, at 3:23 PM, Olivia Guest <notifications@github.com> wrote:
>
>
> OK, cool! @libertyh <https://github.com/libertyh> do you want to review this given the above (no conflict from JOSS' perspective)?
>
