### 2020-10-02 alexrockhill/pd-parser#1 @rougier
Problem with python examples
I've installed the latest version and tried to run the two examples from https://alexrockhill.github.io/pd-parser/auto_examples/index.html but it failed. Not sure about the exact reason (missing data?)

```
$ python plot_find_pd_events.py

Creating RawArray with float64 data, n_channels=1, n_times=2044106
    Range : 0 ... 2044105 =      0.000 ...  2044.105 secs
Ready.
Traceback (most recent call last):
  File "plot_find_pd_events.py", line 33, in <module>
    raw, beh_df, events, corrupted_indices = \
ValueError: not enough values to unpack (expected 4, got 2)
```

```
$ python plot_find_pd_on_and_off.py

  File "plot_find_pd_on_and_off.py", line 40, in <module>
    pd_parser.simulate_pd_data(n_events=n_events, n_secs_on=n_secs_on,
TypeError: simulate_pd_data() got an unexpected keyword argument 'n_secs_on'
```
