### 2020-10-02 alexrockhill/pd-parser#1 @alexrockhill
Sorry the examples are unfortunately not backward compatible and don't work with the latest stable version, only the development version. I'll make a new release but in the meantime, you can run `pip install git+https://github.com/alexrockhill/pd-parser.git`to install the development version but I'll work on adding legacy documentation so this problem doesn't happen again also.
