+++ README.rst
@@ -3,6 +3,8 @@
 |icon| pd-parser
 ================
 
+The development version of the documentation can be found `here <https://alexrockhill.github.io/pd-parser/dev/>`_.
+
 .. |icon| image:: https://alexrockhill.github.io/pd-parser/_images/favicon.png
   :width: 25
 
