+++ docsrc/Makefile
@@ -20,4 +20,4 @@ help:
 	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
 	cp -a figs/* _build/html/_images/
 	cp favicon.png _build/html/_images/
-	cp -a _build/html/. ../docs
+	cp -a _build/html/. ../docs/dev/
