### 2020-10-04 #2674 @oliviaguest
@rougier it's so cute — see: https://joss.readthedocs.io/en/latest/whedon.html
