### 2020-10-05 #2674 @rougier
@guest I can do all these command on ReScience (just replace whedon by rougier)...

@alexrockhill I'started testing your software and opened an issue on your repo. You already addressed one of my issue, could you have a look at the new one?
