### 2020-10-07 alexrockhill/pd-parser#1 @rougier
Thanks. I thought I posted a second issue but I cannot find it. I now got:

```
Used Annotations descriptions: ['Fixation', 'Go Cue', 'Response']
/usr/local/lib/python3.8/site-packages/pd_parser/parse_pd.py:879: RuntimeWarning: The unit for channel(s) pd has changed from V to NA.
  raw.set_channel_types({ch: 'stim' for ch in pd_channels
Traceback (most recent call last):
  File "plot_find_pd_events.py", line 127, in <module>
    pd_parser.pd_parser_save_to_bids(op.join(out_dir, 'bids_dir'), fname,
  File "/usr/local/lib/python3.8/site-packages/pd_parser/parse_pd.py", line 881, in pd_parser_save_to_bids
    bids_path = mne_bids.BIDSPath(subject=sub, session=ses, task=task,
AttributeError: module 'mne_bids' has no attribute 'BIDSPath'
```
