### 2020-10-07 alexrockhill/pd-parser#1 @alexrockhill
Ah, yeah the bleeding edge mne-bids release is used and it's in requirements.txt but not the setup.py dependencies. I'll add to install the requirements (`pip install -r requirements.txt`) until the next `mne-bids` release.
