### 2020-10-07 alexrockhill/pd-parser#1 @rougier
Ok. tell me when it's done such that I can test reinstall with pip.
