### 2020-10-07 alexrockhill/pd-parser#1 @alexrockhill
I changed the README and documentation to say that you need to pip install the bleeding edge mne-bids after pip installing pd-parser. I can update pypi too though I'll let you know when I've done that
