### 2020-10-07 alexrockhill/pd-parser#1 @rougier
That did the trick. Now for the second example, I got:
```
Traceback (most recent call last):
  File "plot_find_pd_on_and_off.py", line 80, in <module>
    pd_parser.add_pd_off_event(fname, off_event_name='Stim Off')
AttributeError: module 'pd_parser' has no attribute 'add_pd_off_event'
```
