### 2020-10-07 alexrockhill/pd-parser#1 @alexrockhill
It looks like I'd have to release a new version so I think I'll just have to leave it as just the documentation changed re the last issue.

Looking at the most recent traceback now.
