+++ setup.py
@@ -74,7 +74,7 @@ if __name__ == "__main__":
           install_requires=[
               'numpy>=1.19.0',
               'mne>=0.19.1',
-              'mne-bids>=0.5',
+              'mne-bids',
               'pybv',
               'matplotlib',
               'argparse',
