+++ docsrc/auto_examples/plot_find_pd_on_and_off.rst
@@ -99,8 +99,8 @@ electrophysiology format)
     Creating RawArray with float64 data, n_channels=3, n_times=2044106
         Range : 0 ... 2044105 =      0.000 ...  2044.105 secs
     Ready.
-    Writing /private/var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_kzljuy96/sub-1_task-mytask_raw.fif
-    Closing /private/var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_kzljuy96/sub-1_task-mytask_raw.fif [done]
+    Writing /private/var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_0w1yc41c/sub-1_task-mytask_raw.fif
+    Closing /private/var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_0w1yc41c/sub-1_task-mytask_raw.fif [done]
 
 
 
@@ -142,17 +142,17 @@ some were too far off from large offsets that we're going to exclude them.
 
  .. code-block:: none
 
-    Reading in /var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_kzljuy96/sub-1_task-mytask_raw.fif
-    Opening raw data file /var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_kzljuy96/sub-1_task-mytask_raw.fif...
+    Reading in /var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_0w1yc41c/sub-1_task-mytask_raw.fif
+    Opening raw data file /var/folders/s4/y1vlkn8d70jfw7s8s03m9p540000gn/T/tmp_mne_tempdir_0w1yc41c/sub-1_task-mytask_raw.fif...
     Isotrak not found
         Range : 0 ... 2044105 =      0.000 ...  2044.105 secs
     Ready.
     Reading 0 ... 2044105  =      0.000 ...  2044.105 secs...
     Finding photodiode events
[…]
