### 2020-10-08 alexrockhill/pd-parser#2 @rly
Provide developer test instructions
There should be documentation on how to run automated tests.
https://joss.readthedocs.io/en/latest/review_criteria.html#tests

Note that one of the tests requires `bids-validator`. It looks like that needs to be installed globally via `npm
install -g bids-validator` (`pip install bids-validator` did not install an executable, at least on my system).

Example documentation:

- Run `pip install pytest`
- Install Node.js (at least version 10.11.0)
- Run `npm install -g bids-validator`
- Run `pytest`
