### 2020-10-08 alexrockhill/pd-parser#3 @rly
Provide contributor guidelines
There should be clear guidelines for third-parties wishing to:
- Contribute to the software
- Report issues or problems with the software
- Seek support

https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines

This is missing.
