### 2020-10-08 alexrockhill/pd-parser#4 @rly
Paper comments
1. Please add your name to the paper.

2. Minor edits:

In summary:
"Events are excluded where when" -> "Events are excluded when"
"for one or more frames, this is" -> "for one or more frames; this is"
"may effect neighboring channels" -> "may affect neighboring channels"
"sychnronize" -> "synchronize"

Figure 2 caption:
"photoiode" -> "photodiode"

Figure 3 caption:
"the difference plot Figure 2" -> "the difference plot, Figure 2"

Statement of need:
"Not only does pd-parser offer" -> change font of "pd-parser"
"serves a forum" -> "serves as a forum"

The code examples could also use a read-through and spelling/grammar check.

3. In your statement of need, you write: "leading to the creation of redundant algorithms that, due to this redundancy, are vulnerable to coding errors."
I disagree. I think redundancy does not necessarily result in vulnerability to coding errors. Rather, a lack of testing over a broad set of inputs and use cases makes the software vulnerable to error.

4. It would be helpful to have a broad statement in the summary about why this tool was developed and what problem it solves. It is mentioned off-hand later: "precision timing of the display is critical, for example vision or psychophysics research", and the documentation includes this line: "This tool is for research that needs high-performance timing synchronization, especially in clinical settings where having a channel record a photodiode is the best or only method of task-integration available and where the photodiode is especially likely to be corrupted by unavoidable environmental noise." This would be helpful context to include in the summary for a diverse, non-specialist audience.

Perhaps something like, "Precise time synchronization between a monitor and a data acquisition system is critical for applications such as vision or psychophysics research. To achieve this, researchers often record voltages from a photodiode attached to a monitor and compute, either manually or automatically, the difference between the times of trigger pulses sent to the monitor and the times of the actual luminance changes from the photodiode. This computation can be difficult when measurements from the photodiode are corrupted or unreliable. pd-parser helps users match photodiode events to timestamped events particularly when the photodiode signal is corrupted."
