### 2020-10-08 alexrockhill/pd-parser#5 @rly
Suggested improvements to example notebooks
1. In the `pd_parser.find_pd_params` GUI:
For the scale step (step 3), it would be nice to show an expanded x axis to see all events. It is difficult to scroll through to find a good value.

2. It might be useful to print to stdout the results of the `pd_parser.find_pd_params` GUI so that they are accessible after closing the GUI.

3. In the `plot_find_pd_on_and_off` notebook, the text says that three events were corrupted, but the plots and output show only two events that were corrupted (= where no event was found). Is one missing?

4. In the `plot_find_pd_on_and_off` notebook, the last plot is hard to read, as mentioned in the notebook. Perhaps a scatter plot comparing the ground truth duration vs the recovered duration may be more effective:

```python
recovered = (off_events[:, 0] - on_events[:, 0]) / raw.info['sfreq']
not_corrupted = [s != 'n/a' for s in beh_df['pd_sample']]
ground_truth_not_corrupted = n_secs_on[not_corrupted]

fig, ax = plt.subplots()
ax.scatter(ground_truth_not_corrupted, recovered)
ax.set_xlabel('ground truth duration')
ax.set_ylabel('recovered duration')
```
