### 2020-10-08 #2674 @rly
Hi @alexrockhill, I have reviewed your software and paper and created a few issues in your repo:
https://github.com/alexrockhill/pd-parser/issues/2
https://github.com/alexrockhill/pd-parser/issues/2
https://github.com/alexrockhill/pd-parser/issues/2
