### 2020-10-08 alexrockhill/pd-parser#4 @alexrockhill
Great, thanks for the issues, I'm almost done with the first two and then I'll add these changes.
