### 2020-10-08 #2674 @rly
Hi @alexrockhill, I have reviewed your software and paper. Overall, the software and paper look good. There were a few issues to address, as per the JOSS paper guidelines/rubric. I created a few issues in your repo:
https://github.com/alexrockhill/pd-parser/issues/2
https://github.com/alexrockhill/pd-parser/issues/3
https://github.com/alexrockhill/pd-parser/issues/4

Suggestions:
https://github.com/alexrockhill/pd-parser/issues/5
