### 2020-10-08 3271d5a
added bug report and contributing guidelines
