### 2020-10-08 alexrockhill/pd-parser#2 @alexrockhill
Ok addressed in 3271d5ab88c08e1e03ad045b9b1f8f5be09f260e, it's very similar to the `mne-bids` testing as that's where the validator tests are from.
