+++ README.rst
@@ -67,7 +67,7 @@ Contributing
 ------------
 
 Please see our `contributing guide <https://github.com/alexrockhill/pd-parser/blob/master/CONTRIBUTING.md>`_.
-
+`
 
 Alex Rockhill
 Eugene, OR 2020
