### 2020-10-08 alexrockhill/pd-parser#3 @alexrockhill
Hmm okay, I did it similar to other projects for the contributing guidelines but I didn't really include anywhere to seek support besides the issue tracker 5462da8d081883b26e1b8e81a65e47a7aa2ef253. I think that's okay for now by me but maybe I will make a Gitter or something soon unless you'd suggest doing that now.
