### 2020-10-08 alexrockhill/pd-parser#4 @alexrockhill
Ok for adding the authors to the paper in the repository, it is added when it's formatted by JOSS but there isn't a simple way from the base command `pandoc -s paper.md --bibliography paper.bib -o paper.pdf` because JOSS does some scripting to make them nicely formatted and the paper.md and paper.bib need to be compatible with JOSS. That might be nice to have at some point but for now I'm just removing the `paper.pdf`.
