### 2020-10-08 0b897af
another paper readthrough and spelling and grammar check
