+++ paper.md
@@ -27,21 +27,21 @@ bibliography: paper.bib
 
 # Summary
 
-Precise synchronization of events displayed on a monitor to recordings of time series data is critical for applications such as vision or psychophysics research. To achieve this, researchers often use a photodiode to convert the luminance on a monitor over time into a voltage time course, which is what is recorded. ``pd-parser`` matches photodiode deflection events to time-stamped events; it is particularly useful when the photodiode signal is corrupted or there is drift between the computer clock controlling the monitor and the data acquisition computer clock. To do this, ``pd-parser`` first extracts photodiode time series data from an electrophysiology data file. ``pd-parser`` uses ``mne-python`` [@Agramfort2013] for input/output. The photodiode data can be on a single channel or two channels that are then bipolar re-referenced. Candidate photodiode events from this photodiode time series are identified based on matching a square-wave template. Time-stamped events from the computer triggering changes to the luminance of the photodiode are read from a tab-separated value (tsv) file, and the best alignment of these events relative to photodiode deflection events is then found while accounting for any drift between computer clocks of separate recording devices like in \autoref{fig:trend}. The success of the alignment can be determined from a plot of event differences, \autoref{fig:diffs}, where when almost all of the events match within around 30 ms, the alignment is assured to be correct as this would basically never happen by chance. Events are excluded when the difference between the photodiode event and the time-stamped event are greater than a specified threshold, as shown in \autoref{fig:exclude}. This discrepancy in timing can occur when a monitor doesn't update the display for one or more frames; this is most often because the computer paused execution of the program controlling luminance to do background tasks. Additional events can then be added using the time of each event relative to the corresponding photodiode-synchronized event. Adding relative events may be needed for any number of reasons. For instance, events may occur too rapidly in series relative to the monitor refresh rate to have a photodiode deflection for each event, or the photodiode deflection may affect neighboring channels on the amplifier. In the latter case, an event that is not analyzed, such as a fixation stimulus, can be chosen to synchronize the photodiode so that other, more important events will not be contaminated by a large photodiode deflection. Finally, the raw data and events data can be saved in brain imaging data structure (BIDS) format, which allows the behavioral events to be stored in a standardized format without modifying the underlying raw electrophysiology file.
[…]
