### 2020-10-08 20b1582
clean up examples
