+++ docs/dev/_downloads/332eafcd55124069bc52ae95a623ffe2/plot_find_pd_events.py
@@ -2,8 +2,8 @@
 ==========================
 01. Find Photodiode Events
 ==========================
-In this example, we use pd-parser to find photodiode events and
-align them to behavior. Then, save the data to BIDS format.
+In this example, we use ``pd-parser`` to find photodiode events and
+align them to behavior. Then, we save the data to BIDS format.
 """
 
 # Authors: Alex Rockhill <aprockhill@mailbox.org>
@@ -13,9 +13,9 @@ align them to behavior. Then, save the data to BIDS format.
 ###############################################################################
 # Simulate data and use it to make a raw object:
 #
-# We'll make an mne.io.Raw object so that we can save out some random
-# data with a photodiode event channel in it in fif format (a standard
-# electrophysiology format)
+# We'll make an `mne.io.Raw object` so that we can save out some random
+# data with a photodiode event channel in it in `fif` format (a commonly used
+# electrophysiology data format).
 import os.path as op
 import numpy as np
 
@@ -26,6 +26,7 @@ import pd_parser
 from pd_parser.parse_pd import _to_tsv
 
 out_dir = _TempDir()
+print(f'After running this example, you can find the data here: {out_dir}')
 
 # simulate photodiode data
 n_events = 300
@@ -79,9 +80,11 @@ _to_tsv(behf, beh_df)
 ###############################################################################
 # Use the interactive graphical user interface (GUI) to find parameters:
 #
-# On the webpage, this is example is not interactive, but if you copy this
-# code into a python console, you can see how to interact with the photo-
-# diode data to pick reasonable parameters by following the instructions.
+# On the documentation webpage, this is example is not interactive,
+# but if you download it as a jupyter notebook and run it or copy the code
+# into a console running python (ipython recommended), you can see how to
+# interact with the photodiode data to pick reasonable parameters by
+# following the instructions.
 
 pd_parser.find_pd_params(fname, pd_ch_names=['pd'])
 
@@ -91,22 +94,25 @@ pd_parser.find_pd_params(fname, pd_ch_names=['pd'])
 # This function will use the default parameters or the parameters you
 # found from :func:`pd_parser.find_pd_parameters` to find and align the
 # photodiode events, excluding events that were off because the commuter
-# hung up on computation for instance. That data is save in the same folder
-# as the raw file which can be used directly or accessed via
-# :func:`pd_parser.pd_parser_save_to_bids`.
+# hung up on computation, for instance. That data is saved in the same folder
+# as the raw file (in this case, a temperary directory generated by
+# :func:`_TempDir`). The data can be used directly, or it can be accessed via
+# :func:`pd_parser.pd_parser_save_to_bids` to store it in the brain imagine
+# data structure (BIDS) standardized format before using it.
 
-pd_parser.parse_pd(fname, behf=behf, pd_ch_names=['pd'])
+pd_parser.parse_pd(fname, behf=behf, pd_ch_names=['pd'], max_len=1.5)
 
 ###############################################################################
 # Add events relative to the photodiode events:
 #
-# The photodiode is usually sychronized to one event (usually the fixation
+# The photodiode is usually sychronized to one event (e.g. the fixation
 # so that if the deflections caused by the photodiode are large enough
 # to influence other channels through amplifier interactions it doesn't
 # cause issues with the analysis) so often the events of interest are
 # relative to the photodiode event. In the task a timer can be started at the
-# photodiode event and pulled for time at each of the following events.
-# These events are then passed in tsv file to be added to the events.
+# photodiode event and checked each time a subsequent event occurs.
+# These events should then be recorded in tsv file, which can be passed to
+# ``pd-parser`` in order to add the events.
[…]

[…]
