### 2020-10-08 alexrockhill/pd-parser#5 @alexrockhill
Great suggestions, I incorporated them in this commit 20b15820ddf56f5f72a636c1504f61240b68efd8
