### 2020-10-09 #2674 @rougier
the two provided examples are now working properly on my machine. I think it would be nice to have these exampls tests alongside the pd_parser such that you can type `pd_parser.test()` to check everything's ok. And/or you could also have very simple unittests that would help you spotting discrepancies in the API. Apart from these minor points, I'm satisfied with the submission.
