### 2020-10-09 9086579
added test for examples
