### 2020-10-09 #2674 @alexrockhill
> the two provided examples are now working properly on my machine. I think it would be nice to have these exampls tests alongside the pd_parser such that you can type `pd_parser.test()` to check everything's ok. And/or you could also have very simple unittests that would help you spotting discrepancies in the API. Apart from these minor points, I'm satisfied with the submission.

Good idea, I added unit tests for the examples to make sure they run with this commit 9086579ba5ca615bc9a2d9e03ede3c406c780073. They don't check whether they look right, the only way to do that is by eye, but you can run `pytest` and it will test everything including the examples now as described in the contributing guidelines.
