### 2020-10-10 alexrockhill/pd-parser#6 @libertyh
Clarification to README.rst
As a quick clarification for those looking for download the examples, I suggest that the README.rst be changed to reflect that the Examples page is part of the Documentation page (and not within the examples directory within the git repo, for example). For example, in "Getting Started", the following text could have a link to https://alexrockhill.github.io/pd-parser/auto_examples/index.html .

> In the 'Examples' page, click on 'Download all examples in Jupyter notebooks'
