### 2020-10-10 alexrockhill/pd-parser#6 @alexrockhill
Good suggestion, I've added it in the latest commit ba121ae165977c61f754ca7fe8ed63e51c9f73b5.

It's only in the development documentation for now but it will be in the stable documentation soon when the next version release happens.
