### 2020-10-10 alexrockhill/pd-parser#6 @alexrockhill
It is in the README now though too.

I wouldn't want someone to try to use it for the first time and hit a roadblock there though... actually I'm changing the version 0.2 documentation too now, great point
