### 2020-10-10 alexrockhill/pd-parser#7 @libertyh
Paper edit suggestions
Thanks for providing this tool. I have a few suggestions for edits to the paper itself. Some of these suggestions are made perhaps because I myself don't typically use a photodiode to synchronize data (we use audio + a match filter instead), but I think overall this will be helpful for those contemplating using this type of setup and knowing that such a tool is available.

1. In the **"Statement of Need"** section, you mention that for your EMU recordings, "a photodiode placed on the monitor of the laptop was used to detect task-related luminance changes. The photodiode recordings were then digitized by the same clinical recording system which was used for the electrophysiology recordings." I'm not familiar with exactly how this is typically done. Do you mean that the photodiode itself is used as an additional channel in the preamplifier/headbox? Just adding a short sentence about how this is used with your system would be helpful.

2. If this is not appropriate for JOSS papers, feel free to ignore, but I felt the paper could have a brief statement of what general algorithm was used for event detection. On page 3, the authors say "This package has robust photodiode event-determination and photodiode to time-stamp alignment algorithms that are validated with both real and simulated photodiode data", and then go on to describe the possible differences in experimental setups. I think it would be helpful to add that this detects square-wave like deflections based on their width and a z-score threshold (or some other general description of the process). This helps me to know what the limitations might be and why some events may not be detected adequately.

3. Related to point 2, in Figure 3, the first row, third column says "no event found", but is this a false negative? Is this because of a short jump in the voltage that should have been ignored? I see in the legend this is discussed somewhat. Perhaps also a report on the total number of events expected and events detected could give the user an idea of the sensitivity/specificity of detection based on the parameters chosen.
