### 2020-10-10 alexrockhill/pd-parser#7 @alexrockhill
I'll incorporate the suggestions into the paper, thanks!

We had a redundant setup at first with audio as well and I think the fidelity is better (i.e. not having corrupted events) but for timing related to screen events I think the latency and uncertainty with the photodiode is better (for button or other events the audio is probably the same). Even with state-of-the-art setups and refresh rates, I noticed, especially working in vision research, that the monitor could hang a frame or more up to 5-10% of the time especially with older monitors and computers.

I was thinking of adding an audio match for the next release though because that is definitely widely used too.
