### 2020-10-10 #2674 @libertyh
I've also started my review and left some comments about the README and the paper text so far. @oliviaguest I think I may need to be re-invited to the repo (sorry!!) in order to complete the checklist. Will link to my issues in a bit when complete!
