### 2020-10-10 alexrockhill/pd-parser#7 @alexrockhill
To answer your question about how we record the data, yes, the photodiode is plugged into the headstage/pre-amp through a gain modifier and bipolar referenced. You could try one channel with a common reference and you might be able to get away with it but we were having corrupted events even with the bipolar setup, although as you can see in the example recording the baseline noise was low, it was just there was a fair amount of drift and a lot of artifacts.
