### 2020-10-10 alexrockhill/pd-parser#7 @alexrockhill
Also for point 2 and 3 and why there was no event found for 98 (and this is related to example 2) the algorithm requires a square wave with a stretch of data greater than the zscore, and then an offset without any reonset. In this case, the algorithm rejected what could be considered an event because of the reonset. Example 2, that I just mentioned, shows getting the offset events, which would be complicated and I thought unideal if for every onset event you didn't have a corresponding offset event that was clean.
