### 2020-10-10 alexrockhill/pd-parser#7 @libertyh
> To answer your question about how we record the data, yes, the photodiode is plugged into the headstage/pre-amp through a gain modifier and bipolar referenced. You could try one channel with a common reference and you might be able to get away with it but we were having corrupted events even with the bipolar setup, although as you can see in the example recording the baseline noise was low, it was just there was a fair amount of drift and a lot of artifacts.

Thanks! That helps -- if you want this to be adopted mostly by SEEG/ECoG people I could see this being a helpful detail, but I know photodiode event detection is widely used in other fields, so I'm also not sure whether this level of detail would be needed in the paper. Maybe it can go in the docs or somewhere instead, if anywhere.
