### 2020-10-10 alexrockhill/pd-parser#7 @alexrockhill
> > Also for point 2 and 3 and why there was no event found for 98 (and this is related to example 2) the algorithm requires a square wave with a stretch of data greater than the zscore, and then an offset without any reonset. In this case, the algorithm rejected what could be considered an event because of the reonset. Example 2, that I just mentioned, shows getting the offset events, which would be complicated and I thought unideal if for every onset event you didn't have a corresponding offset event that was clean.
>
> That's a good point. I think like a lot of things then, it requires the user to look through the data to determine what works best for their particular application, and how many of those events they'd want to remove vs. retain.

Yeah events like those could be added back but I think this was some of the most corrupted photodiode data I've seen (shown in the example) and it still only lost ~3% of trials with the strictest time lag constraints (the 37 and 43 ms differences could easily have been kept). I think for most situations that's an acceptable loss and if you're having artifacts in the photodiode on a trial, that might be during the trial and so it might be a good trial to exclude just to be safe anyway.

The example data was collected with a photodiode fixed to the screen with adhesive and plastic supports but on a laptop on the participant's lap who moves around sometimes so I think that is pretty much the worst situation for photodiode artifact.
