+++ paper.md
@@ -33,13 +33,13 @@ Precise synchronization of events displayed on a monitor to recordings of time s
 
 ![*An Example of Event Differences*. An example of the differences between the photodiode deflection event and the time-stamped event aligned to it, excluding events that were not matched or that were matched but the difference was too much, as shown in \autoref{fig:exclude}. As can be observed in the figure, the alignments are much closer than could have possibly happened by chance, giving confidence that the fit of the events was correct. \label{fig:diffs}](docsrc/figs/event_diffs.png)
 
-![*An Example of Excluded Events*. An example of the excluded events from 300 time-stamped events synchronized to photodiode deflection events is shown. The threshold of exclusion can be set to include all the events except the those where no event was found, but from inspecting \autoref{fig:diffs}, the difference plot, these event differences are subjectively outliers. \label{fig:exclude}](docsrc/figs/excluded_events.png)
+![*An Example of Excluded Events*. An example of the excluded events from 300 time-stamped events synchronized to photodiode deflection events is shown. The threshold of exclusion can be set to include all the events except the those where no event was found, but from inspecting \autoref{fig:diffs}, the difference plot, these event differences are subjectively outliers. Trials where no event was found can be recovered manually when the pre-event baseline, plateau or post-event baseline was corrupted but photodiode onset or offset was not. \label{fig:exclude}](docsrc/figs/excluded_events.png)
 
 
 
 # Statement of need 
 
-To our knowledge, there are no software packages that extract photodiode events and align them with event time-stamps, despite the widespread use of photodiodes for task timing. By sensing luminance, photodiodes can synchronize recording systems with high temporal precision. For example, the timing of events in behavioral tasks displayed on a laptop can be synchronized to electrophysiological recordings. While many research systems are specially designed to record triggers that link recordings from separate machines directly, other recording systems, especially clinical systems, lack this capability. In these cases, use of a photodiode offers a robust and reliable method for synchronization. We developed this software package to address our need to synchronize task events with intracranial recordings acquired in the epilepsy monitoring unit. Here electrophysiology was acquired using a clinical system and a behavioral task was performed using a laptop brought to the patient’s bedside. A photodiode placed on the monitor of the laptop was used to detect task-related luminance changes. The photodiode recordings were then digitized by the same clinical recording system which was used for the electrophysiology recordings. Due to variability in refresh rates for monitors, the use of photodiodes is especially helpful for research where precision timing of the display is critical, for example vision or psychophysics research. This software package addresses photodiode synchronization in a comprehensive way so that photodiode parsing can be done by flexibly changing key parameters, which avoids writing entirely new scripts. This reduces the inefficiency caused by needing to make an algorithm for each particular situation, and it reduces the potential for errors by testing a broad array of situations thoroughly.
[…]
