+++ pd_parser/parse_pd.py
@@ -253,15 +253,27 @@ def _find_best_alignment(beh_events, sorted_pds, max_i, sfreq,
     return best_alignment
 
 
+def _recover_event(pd, b_event, exclude_shift_i, zscore):
+    """Recover with a corrupted baseline or plateau but not on/offset."""
+    b_event_i = np.round(b_event).astype(int)
+    section = pd[b_event_i - exclude_shift_i:
+                 b_event_i + exclude_shift_i].copy()
+    baseline = pd[b_event_i - 2 * exclude_shift_i: b_event_i - exclude_shift_i]
+    section = (section - np.median(baseline)) / baseline.std()
+    event = np.where(abs(section) > zscore)[0]
+    return b_event_i - exclude_shift_i + event[0]if event.size > 0 else None
+
+
 def _exclude_ambiguous_events(beh_events, sorted_pds, best_alignment, pd,
                               exclude_shift_i, max_len_i, resync_i, sfreq,
-                              verbose=True):
+                              zscore, recover, verbose=True):
     """Exclude all events that are outside the given shift compared to beh."""
     if verbose:
         import matplotlib.pyplot as plt
         pd_section_data = dict(b_event=list(), title=list())
         print('Excluding events that have zero close events or more than '
               'one photodiode event within `max_len` time')
+    sfreq_i = np.round(sfreq).astype(int)
     events = dict()
     errors = dict()
     pd_candidates = set(sorted_pds)
@@ -295,7 +307,28 @@ def _exclude_ambiguous_events(beh_events, sorted_pds, best_alignment, pd,
                     j_ms = int(j / sfreq * 1000)
                     text = f'Excluding event {i},\noff by {j_ms} ms'
                 else:  # if off by more than max_len, just say missing
-                    text = f'Excluding event {i},\nno event found'
+                    if recover:
+                        event = _recover_event(pd, b_event, exclude_shift_i,
+                                               zscore)
+                        if event is None:
+                            text = f'No event found to recover\nfor event {i}'
+                        else:
+                            fig, ax = plt.subplots()
+                            section = \
+                                pd[event - 2 * sfreq_i: event + 2 * sfreq_i]
+                            ax.plot(np.linspace(-2, 2, 4 * sfreq_i), section)
+                            ax.plot([0, 0], [section.min(), section.max()])
+                            ax.set_title(f'Corrupted Event {i}')
+                            ax.set_xlabel('time (s)')
+                            ax.set_ylabel('voltage')
+                            fig.show()
+                            if input('Recover event? (y/N) ').lower() == 'y':
+                                events[i] = event
+                                text = f'Event {i} recovered\n(not excluded)'
+                            else:
+                                text = f'Recovered event {i}\ndiscarded'
+                    else:
+                        text = f'Excluding event {i},\nno event found'
                 print(text.replace('\n', ' '))
                 pd_section_data['b_event'].append(b_event)
                 pd_section_data['title'].append(text)
@@ -552,7 +585,7 @@ def parse_pd(fname, pd_event_name='Fixation', behf=None,
              beh_col='fix_onset_time', pd_ch_names=None,
              exclude_shift=0.03, resync=0.075,
              max_len=1., zscore=100, min_i=10, baseline=0.25,
-             add_events=False, overwrite=False, verbose=True):
+             add_events=False, recover=False, overwrite=False, verbose=True):
     """Parse photodiode events.
 
     Parses photodiode events from a likely very corrupted channel
[…]

[…]
