### 2020-10-11 alexrockhill/pd-parser#7 @alexrockhill
I thought about what you said Liberty and I thought it could be helpful to add the ability to recover events that still look good. Here's the example I wrote https://alexrockhill.github.io/pd-parser/dev/auto_examples/plot_recover_events.html#sphx-glr-auto-examples-plot-recover-events-py.

Note that it's interactive (just pass/fail input for onset but you have to find the event manually for offset) so you have to run it to see that part.
