### 2020-10-12 #2674 @rougier
Good bot ! I forgot to check one box and I just remember I had a question about co-authors. They do not seem to have commited anything in the repo. @alexrockhill Could you explain their role ?
