### 2020-10-12 #2674 @alexrockhill
> Good bot ! I forgot to check one box and I just remember I had a question about co-authors. They do not seem to have commited anything in the repo. @alexrockhill Could you explain their role ?

Dr. Raslan and Dr. Swann have active supervision, review and high-level direction of both the software package and the physical setup for data collection that is used in the example data and discussed in the paper.
