### 2020-10-12 #2674 @libertyh
Hi everyone! Thanks @alexrockhill for starting to address my comments. I'm including the issues I opened here:

alexrockhill/pd-parser#6
alexrockhill/pd-parser#7
alexrockhill/pd-parser#8

Most of these have to do with clarifications to the usage of pd-parser so that it's as easy as possible for your end users to get started. I was able to get the code to work on my machine with the jupyter notebook examples. Happy to discuss when everyone is ready.
