### 2020-10-12 alexrockhill/pd-parser#8 @alexrockhill
I would note that the photodiode channel (in reference to `find_pd_params` and other functions) is an optional argument and if you don't know it when you call the function, the data will be plotted and user input will be requested.
