### 2020-10-12 alexrockhill/pd-parser#8 @alexrockhill
I incorporated your suggestions in d98990fe0291bfbaa015aee3852c79b907eaa490, thanks!
