### 2020-10-22 #2674 @rly
Everything looks good on my end. The new example mentioned in https://github.com/alexrockhill/pd-parser/issues/7 looks good too. @alexrockhill I would encourage you to git tag the new version and make a new release both on PyPI and on GitHub.

Pending that, I think that the tool and paper are both suitable for publication on JOSS.
