### 2020-10-22 #2674 @libertyh
Looks good to me too, I agree with @rly that following the new release this is acceptable for publication in JOSS.
