+++ docs/_sources/auto_examples/sg_execution_times.rst.txt
@@ -5,10 +5,12 @@
 
 Computation times
 =================
-**00:43.053** total execution time for **auto_examples** files:
+**00:46.944** total execution time for **auto_examples** files:
 
 +-------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_plot_find_pd_events.py` (``plot_find_pd_events.py``)         | 00:43.053 | 0.0 MB |
+| :ref:`sphx_glr_auto_examples_plot_find_pd_events.py` (``plot_find_pd_events.py``)         | 00:20.175 | 0.0 MB |
 +-------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_plot_find_pd_on_and_off.py` (``plot_find_pd_on_and_off.py``) | 00:00.000 | 0.0 MB |
+| :ref:`sphx_glr_auto_examples_plot_find_pd_on_and_off.py` (``plot_find_pd_on_and_off.py``) | 00:14.849 | 0.0 MB |
++-------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_auto_examples_plot_recover_events.py` (``plot_recover_events.py``)         | 00:11.919 | 0.0 MB |
 +-------------------------------------------------------------------------------------------+-----------+--------+
