### 2020-10-23 #2674 @alexrockhill
Here's the DOI for the 0.3 release as well http://doi.org/10.5281/zenodo.4122049
