### 2020-10-25 #2674 @kyleniemeyer
@alexrockhill Hi there, I am the EIC on duty this week doing some final checks before publishing. There are just a few things I noticed:

- Can you break up the text of your article into paragraphs? Right now it sort of looks like a wall of text in each section.
- Can you add your co-authors to the Zenodo archive? That author list should match the paper as well.
