### 2020-10-25 #2674 @alexrockhill
Thanks for the copy editing, @kyleniemeyer, those suggested changes have been implemented!
